feature: create imageInput FilseDrop and ImageCropp
